pyproject.toml
[build-system]
build-backend = "setuptools.build_meta"
requires = ["setuptools>=61.0"]

[project]
name = "sthali-cli"
version = "0.0.0.1"
authors = [{ name = "Jhunu Fernandes", email = "jhunu.fernandes@gmail.com" }]
description = "A CLI package for Sthali development"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  # "Programming Language :: Python :: 3.11",
  # "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Internet",
  "Topic :: Software Development :: Libraries :: Application Frameworks",
  "Topic :: Software Development :: Libraries :: Python Modules",
  "Topic :: Software Development :: Libraries",
  "Topic :: Software Development",
  "Typing :: Typed",
]
dependencies = [
  "cookiecutter >= 2.6.0",
  "mkdocs >= 1.6.0",
  "PyYAML >= 6.0.2",
  "tomli >= 2.0.1",
  "typer >= 0.12.5",
]

[project.urls]
"Homepage" = "https://github.com/jhunufernandes/sthali-cli"
"Bug Tracker" = "https://github.com/jhunufernandes/sthali-cli/issues"

[project.scripts]
sthali_cli = "sthali_cli:app"

[tool.ruff]
line-length = 119

[tool.ruff.format]
exclude = [
  "{{cookiecutter.project_name}}/",
]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN401",  # any-type
  "TID252",  # relative-imports
  # "ANN002",  # missing-type-args
  # "ANN003",  # missing-type-kwargs
  # "ANN101",  # missing-type-self
  # "ANN102",  # missing-type-cls
  # "COM812",  # missing-trailing-comma
  # "ISC001",  # single-line-implicit-string-concatenation
  # "D107",  # undocumented-public-init
]

[tool.ruff.lint.per-file-ignores]
"src/sthali_db/clients/tinydb.py" = [
  "PGH003",
]
"tests/*" = [
  "COM812",
  "D100",
  "D101",
  "D102",
  "D104",
  "D106",
  "PGH003",
  "PIE804",
  "PT009",
  "PT027",
  "S101",
]

[tool.ruff.lint.pydocstyle]
convention = "google"
